feat: 일부 에셋 코드 에셋 파일 화
