--- components/loading/index.tsx
@@ -1,49 +1,0 @@
-import ManyTrees from '../svgs/many-trees'
-
-const Loading = () => {
-  return (
-    <div className="h-calc-h flex flex-col px-5 py-4 ">
-      <section className="flex flex-col flex-1 justify-center">
-        <h2 className="text-mainTitle2-medium">남이 써주는</h2>
-        <h1 className="text-[40px] font-bold mb-2">나의 소개서</h1>
-        <svg
-          width="182"
-          height="51"
-          viewBox="0 0 182 51"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M180.822 45.3394C180.822 48.0851 178.596 50.3109 175.85 50.3109C173.105 50.3109 170.879 48.0851 170.879 45.3394V21.6277C170.879 18.8821 173.105 16.6562 175.85 16.6562C178.596 16.6562 180.822 18.8821 180.822 21.6277V45.3394Z"
-            fill="#00BC68"
-          />
-          <path
-            d="M145.682 50.9743C143.15 50.9743 140.916 50.3638 138.979 49.1428C137.043 47.9599 135.535 46.2428 134.455 43.9916C133.412 41.7403 132.891 39.0884 132.891 36.0358V21.6043C132.891 18.8586 135.116 16.6328 137.862 16.6328C140.608 16.6328 142.834 18.8586 142.834 21.6043V34.3187C142.834 36.8752 143.318 38.8403 144.286 40.214C145.291 41.5495 146.725 42.2172 148.587 42.2172C150.524 42.2172 152.05 41.4732 153.168 39.985C154.285 38.4969 154.843 36.4555 154.843 33.8608V21.6043C154.843 18.8586 157.069 16.6328 159.815 16.6328C162.561 16.6328 164.786 18.8586 164.786 21.6043V45.2015C164.786 47.9472 162.561 50.173 159.815 50.173H155.685C155.22 50.173 154.843 49.7963 154.843 49.3317C154.843 48.5163 153.707 48.1096 153.076 48.6268C151.169 50.1918 148.704 50.9743 145.682 50.9743Z"
-            fill="#111111"
-          />
-          <path
-            d="M76.9062 21.6283C76.9062 18.8826 79.1321 16.6568 81.8777 16.6568H86.0596C86.4957 16.6568 86.8492 17.0103 86.8492 17.4464C86.8492 18.2099 87.89 18.5757 88.4607 18.0685C89.027 17.5651 89.6256 17.1518 90.2566 16.8285C91.56 16.1798 93.1055 15.8555 94.893 15.8555C96.9412 15.8555 98.8031 16.2752 100.479 17.1147C102.935 18.2891 106.381 18.382 108.833 17.1988C108.934 17.1504 109.035 17.1032 109.137 17.0574C110.925 16.2561 113.047 15.8555 115.505 15.8555C118 15.8555 120.197 16.466 122.097 17.687C123.996 18.8699 125.467 20.587 126.509 22.8382C127.552 25.0895 128.073 27.7415 128.073 30.794V45.2255C128.073 47.9712 125.848 50.197 123.102 50.197C120.356 50.197 118.131 47.9712 118.131 45.2255V32.5111C118.131 29.9546 117.665 28.0086 116.734 26.6731C115.803 25.2994 114.425 24.6126 112.6 24.6126C110.999 24.6126 109.752 25.2422 108.858 26.5013C108.001 27.7605 107.536 29.5921 107.461 31.996V45.2255C107.461 47.9712 105.236 50.197 102.49 50.197C99.7442 50.197 97.5184 47.9712 97.5184 45.2255V32.5111C97.5184 29.9546 97.0529 28.0086 96.1219 26.6731C95.1909 25.2994 93.813 24.6126 91.9883 24.6126C90.387 24.6126 89.1395 25.2422 88.2457 26.5013C87.3892 27.7605 86.9237 29.5921 86.8492 31.996V45.2255C86.8492 47.9712 84.6234 50.197 81.8777 50.197C79.1321 50.197 76.9062 47.9712 76.9062 45.2255V21.6283Z"
-            fill="#00BC68"
-          />
-          <path
-            d="M0 21.6283C0 18.8826 2.22581 16.6568 4.97149 16.6568H9.10163C9.56629 16.6568 9.94298 17.0335 9.94298 17.4981C9.94298 18.3135 11.0796 18.7202 11.71 18.203C13.6173 16.638 16.082 15.8555 19.1039 15.8555C21.6362 15.8555 23.8706 16.466 25.8071 17.687C27.7435 18.8699 29.2331 20.587 30.2758 22.8382C31.3558 25.0895 31.8957 27.7415 31.8957 30.794V45.2255C31.8957 47.9712 29.6699 50.197 26.9242 50.197C24.1786 50.197 21.9528 47.9712 21.9528 45.2255V32.5111C21.9528 29.9546 21.45 28.0086 20.4446 26.6731C19.4763 25.2994 18.0612 24.6126 16.1992 24.6126C14.2628 24.6126 12.736 25.3566 11.6188 26.8448C10.5016 28.3329 9.94298 30.3743 9.94298 32.969V45.2255C9.94298 47.9712 7.71717 50.197 4.97149 50.197C2.22581 50.197 0 47.9712 0 45.2255V21.6283Z"
-            fill="#111111"
-          />
-          <path
-            d="M180.211 2.1858C179.057 1.03589 177.605 0.460938 175.854 0.460938C174.104 0.460938 172.652 1.03589 171.497 2.1858C170.343 3.3357 169.766 4.78235 169.766 6.52576C169.766 8.26916 170.343 9.71582 171.497 10.8657C172.652 12.0156 174.104 12.5906 175.854 12.5906C177.605 12.5906 179.057 12.0156 180.211 10.8657C181.366 9.71582 181.943 8.26916 181.943 6.52576C181.943 4.78235 181.366 3.3357 180.211 2.1858Z"
-            fill="#00BC68"
-          />
-          <path
-            d="M51.4707 50.9983C48.3798 50.9983 45.6241 50.2352 43.2035 48.7089C40.7829 47.2207 38.9023 45.1412 37.5617 42.4701C36.2211 39.7991 35.5508 36.7847 35.5508 33.4269C35.5508 30.069 36.2211 27.0546 37.5617 24.3836C38.9023 21.7126 40.7829 19.633 43.2035 18.1449C45.6241 16.6186 48.3798 15.8555 51.4707 15.8555C54.6079 15.8555 57.1449 16.6402 59.0817 18.2096C59.7152 18.7229 60.8551 18.3135 60.8551 17.4981C60.8551 17.0335 61.2318 16.6568 61.6964 16.6568H65.8266C68.5723 16.6568 70.7981 18.8826 70.7981 21.6283V45.2255C70.7981 47.9712 68.5723 50.197 65.8266 50.197H61.6964C61.2318 50.197 60.8551 49.8203 60.8551 49.3556C60.8551 48.5403 59.7152 48.1309 59.0817 48.6442C57.1449 50.2136 54.6079 50.9983 51.4707 50.9983ZM53.0348 42.2984C55.4554 42.2984 57.3918 41.4781 58.8442 39.8373C60.2965 38.1584 61.0227 36.0216 61.0227 33.4269C61.0227 30.8322 60.2965 28.7145 58.8442 27.0737C57.3918 25.3948 55.4554 24.5553 53.0348 24.5553C50.6887 24.5553 48.8081 25.3948 47.393 27.0737C45.9779 28.7145 45.2703 30.8322 45.2703 33.4269C45.2703 36.0216 45.9779 38.1584 47.393 39.8373C48.8081 41.4781 50.6887 42.2984 53.0348 42.2984Z"
-            fill="#00BC68"
-          />
-        </svg>
-      </section>
-      <div className="ml-5 mt-5"
-      >
-      <ManyTrees /></div>
-    </div>
-  )
-}
-
-export default Loading
